AutogradQNode moved into its own module.
